Cleanup by componentizing elements and styles
