+++ app/components/filters/StatusFilter.tsx
@@ -1,6 +1,6 @@
 import { Checkbox } from "@mantine/core";
 import React from "react";
-import styles from "../../page.module.css";
+import styles from "./FilterSection.module.css";
 import { Status } from "../../global/interfaces";
 
 interface StatusFilterProps {
