+++ app/components/navigation/HeaderSection.module.css
@@ -1,13 +1,25 @@
-html, body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
+.headerContainer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
 }
 
-.wrapper {
+.logoContainer {
+  flex: 0 0 20%;
+  max-width: 250px;
+  justify-content: flex-start;
+  align-items: center;
+  padding: 1rem;
+}
+
+.aboutContainer {
   display: flex;
-  height: 80vh;
-  width: 95vw;
+  justify-content: flex-end;
+  align-items: end;
+  width: 100%;
+  overflow: hidden;
+  text-align: end;
 }
 
 .logo {
@@ -15,22 +27,13 @@ html, body {
   cursor: pointer;
 }
 
-.aboutContainer {
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-  overflow: hidden;
-  text-align: end;
-}
-
 .aboutButton {
   text-transform: uppercase;
-  margin: .8rem 2.2rem .8rem 0;
+  margin: 0.8rem 2.2rem 0.8rem 0;
   cursor: pointer;
   font-size: smaller;
   font-weight: bold;
   border: none;
   text-align: end;
   background-color: transparent;
-    position: absolute;
 }
